Iffy commit
Client: just changed the variable name of the data connection. Server:
Fixed socket error, implemented the while loop that will read each
command and made sure it goes to the right error checking. Also make
sure all replies are being sent to the control connection. Created the
second data connection on the server that will connect to the data
(welcoming) connection on the client. Just need to figure out how to
transfer the file (client and server).